chore: changes per Copilot comments
